# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentCordon API
  version: 1.6.1
  description: |
    AgentCordon is a self-hostable Agentic Identity Provider (IdP) that provides
    secure, auditable, policy-controlled token brokerage for autonomous agents.
    Agents do not directly hold long-lived credentials; every issuance and use
    is observable via the audit log.
  contact:
    name: AgentCordon
  license:
    name: MIT
servers:
  - url: http://localhost:3140
    description: Local development server
...
security:
  - SessionCookie: []
  - BearerJWT: []
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: agtcrdn_session
      description: |
        Session cookie set after user login. Every unsafe method authenticated
        this way is also CSRF-protected by a double-submit rule: the request
        must carry an `X-CSRF-Token` header whose value matches the readable
        `agtcrdn_csrf` cookie, or it is refused with
        `403 csrf_validation_failed`.
    BearerJWT:
      type: http
      scheme: bearer
      description: |
        An **opaque** OAuth 2.0 access token issued to a workspace by
        `POST /api/v1/oauth/token`. Despite the scheme name this is not a JWT
        and carries no claims a client can read: the server stores only its
        SHA-256 hash and resolves the workspace by token id. The ES256 JWT
        issuer, its JWKS endpoint, and the permissions-token endpoint were
        removed.
...
paths:
...
  /api/v1/policies/rsop:
    post:
      tags: [Policies]
      summary: Resultant Set of Policy (RSoP)
      description: |
        Evaluates all active policies against every principal (agents and devices)
        for a given resource, producing a permission matrix. Shows who can do what
        on a specific credential or MCP server. Requires manage_policies permission.
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_type, resource_id]
              properties:
                resource_type:
                  type: string
                  enum: [Credential, McpServer]
                  description: Type of resource to evaluate policies against
                resource_id:
                  type: string
                  format: uuid
                  description: ID of the specific resource to evaluate
                limit:
                  type: integer
                  default: 100
                  description: Maximum number of principals to evaluate
            example:
              resource_type: Credential
              resource_id: "550e8400-e29b-41d4-a716-446655440000"
              limit: 100
      responses:
        '200':
          description: Permission matrix for all principals against the resource
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                  data:
                    type: object
                    properties:
                      resource:
                        type: object
                        properties:
                          type:
                            type: string
                          id:
                            type: string
                            format: uuid
                          name:
                            type: string
                          service:
                            type: string
                      evaluated_at:
                        type: string
                        format: date-time
                      principal_count:
                        type: integer
                      matrix:
                        type: array
                        items:
                          type: object
                          properties:
                            principal_type:
                              type: string
                              enum: [Agent, Device]
                            principal_id:
                              type: string
                              format: uuid
                            principal_name:
                              type: string
                            principal_tags:
                              type: array
                              items:
                                type: string
                            results:
                              type: object
                              description: Map of action name to decision object
                              additionalProperties:
                                type: object
                                properties:
                                  decision:
                                    type: string
                                    enum: [allow, deny]
                                  reasons:
                                    type: array
                                    items:
                                      type: string
                      conditional_policies:
                        type: array
                        items:
                          type: object
                          properties:
                            policy_name:
                              type: string
                            policy_id:
                              type: string
                              format: uuid
                            condition_type:
                              type: string
                            description:
                              type: string
        '400':
          description: Invalid resource_type or missing required fields
        '403':
          description: Access denied by policy
        '404':
          description: Resource not found
...